Put each PID controller into its own source file.
