+++ src/test/Makefile
@@ -609,6 +609,30 @@ $(OBJECT_DIR)/flight/pid.o : \
 	@mkdir -p $(dir $@)
 	$(CC) $(C_FLAGS) $(TEST_CFLAGS) -DBLACKBOX -c $(USER_DIR)/flight/pid.c -o $@
 
+$(OBJECT_DIR)/flight/pid_luxfloat.o : \
+	$(USER_DIR)/flight/pid_luxfloat.c \
+	$(USER_DIR)/flight/pid.h \
+	$(GTEST_HEADERS)
+
+	@mkdir -p $(dir $@)
+	$(CC) $(C_FLAGS) $(TEST_CFLAGS) -DBLACKBOX -c $(USER_DIR)/flight/pid_luxfloat.c -o $@
+
+$(OBJECT_DIR)/flight/pid_mwrewrite.o : \
+	$(USER_DIR)/flight/pid_mwrewrite.c \
+	$(USER_DIR)/flight/pid.h \
+	$(GTEST_HEADERS)
+
+	@mkdir -p $(dir $@)
+	$(CC) $(C_FLAGS) $(TEST_CFLAGS) -DBLACKBOX -c $(USER_DIR)/flight/pid_mwrewrite.c -o $@
+
+$(OBJECT_DIR)/flight/pid_mw23.o : \
+	$(USER_DIR)/flight/pid_mw23.c \
+	$(USER_DIR)/flight/pid.h \
+	$(GTEST_HEADERS)
+
+	@mkdir -p $(dir $@)
+	$(CC) $(C_FLAGS) $(TEST_CFLAGS) -DBLACKBOX -c $(USER_DIR)/flight/pid_mw23.c -o $@
+
 $(OBJECT_DIR)/pid_unittest.o : \
 	$(TEST_DIR)/pid_unittest.cc \
 	$(USER_DIR)/flight/pid.h \
@@ -621,6 +645,9 @@ $(OBJECT_DIR)/pid_unittest : \
 	$(OBJECT_DIR)/common/maths.o \
 	$(OBJECT_DIR)/common/filter.o \
 	$(OBJECT_DIR)/flight/pid.o \
+	$(OBJECT_DIR)/flight/pid_luxfloat.o \
+	$(OBJECT_DIR)/flight/pid_mwrewrite.o \
+	$(OBJECT_DIR)/flight/pid_mw23.o \
 	$(OBJECT_DIR)/pid_unittest.o \
 	$(OBJECT_DIR)/gtest_main.a
